requirements.txt
streamlit
langchain-openai
langchain-core
langchain-prompts
langchain-chroma
openai
